refactor: moved tests back next to source code
