+++ src/graphql/graphql.module.ts
@@ -1,5 +1,4 @@
 import { Module } from '@nestjs/common';
-import MiddlewareModule from './middleware/middleware.module';
 import { GraphQLModule } from '@nestjs/graphql';
 import { ApolloDriver, ApolloDriverConfig } from '@nestjs/apollo';
 import { ApolloServerPluginLandingPageLocalDefault } from '@apollo/server/plugin/landingPage/default';
@@ -13,7 +12,6 @@ import { ApolloServerPluginLandingPageLocalDefault } from '@apollo/server/plugin
       playground: false,
       plugins: [ApolloServerPluginLandingPageLocalDefault()],
     }),
-    MiddlewareModule,
   ],
 })
 export default class GraphqlModule {}
