--- src/graphql/middleware/middleware.module.ts
@@ -1,4 +1,0 @@
-import { Module } from '@nestjs/common';
-
-@Module({})
-export default class MiddlewareModule {}
